Changed definition of users table.
Removed id column from users and added unique constraints to user and
email columns. as these are both required to be individually unique
either could be used as the primary key instead of the synthetic id
value. This should reduce errors when linking tables and prevent
accidental insertion of duplicate values where thy should not be allowed